Reports Millennium item status values in TRANSCODE_HOLD.TRANSCODE
https://trello.com/c/E1A0lX2u

https://tlcprojects.basecamphq.com/projects/13463793-carl-x-implementation-nashville-metro-schools-nashville-public/posts/101120733/comments#comment_366072920

Wrede requests:

In the TRANSITEM_HOLDS.txt file - there appears to have been an attempt to translate
some of the status codes to CARL status codes. Our migration scripts are set to
translate the III status codes to the CARL status codes. Please remove this
translation and continue to provide to us as in the previous extracts.